+++ packages/div-flo-core/src/index.ts
@@ -1,3 +1,7 @@
 // Core package exports
 export { VersionService } from './services/VersionService';
 export { VersionRepository } from './repositories/VersionRepository';
+export { CaptureService } from './services/CaptureService';
+export { UserService } from './services/UserService';
+export { CaptureRepository } from './repositories/CaptureRepository';
+export { UserRepository } from './repositories/UserRepository';