+++ packages/div-flo-models/src/index.ts
@@ -1,6 +1,11 @@
 // DTOs
-export * from './dtos/VersionInfo';
+export * from './dto/VersionInfo';
 
 // Interfaces  
 export * from './interfaces/IVersionService';
 export * from './interfaces/IVersionRepository';
+export * from './interfaces/ICaptureService';
+export * from './interfaces/ICaptureRepository';
+export * from './interfaces/IUserService';
+export * from './interfaces/IUserRepository';
+