+++ CHANGELOG.md
@@ -2,6 +2,36 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.1.11](https://github.com/jgsteeler/divergent-flow-core/compare/v0.1.10...v0.1.11) (2025-10-17)
+
+
+### Features
+
+* add PostgreSQL database configuration and initial migration for User and Capture models ([a085e37](https://github.com/jgsteeler/divergent-flow-core/commit/a085e37884763e983daa0376b4b18d3a36f42bcf))
+* add Prisma client generation steps for div-flo-models and div-flo-core in CI workflow ([990a737](https://github.com/jgsteeler/divergent-flow-core/commit/990a737063248a7213a67961d70289e521a981f9))
+* **api:** Implement Capture and User controllers with CRUD operations ([da3568a](https://github.com/jgsteeler/divergent-flow-core/commit/da3568ac19db57ae0b6153e62e2a425f225669a3))
+* implement Capture and User repositories with integration tests, update package configurations, and add CI workflow ([75f0ad4](https://github.com/jgsteeler/divergent-flow-core/commit/75f0ad4ff7df1fd855fcdc1da6709062606e540d))
+
+### [0.1.10](https://github.com/jgsteeler/divergent-flow-core/compare/v0.1.9...v0.1.10) (2025-10-13)
+
+
+### Features
+
+* implement API versioning and update related routes, tests, and configurations ([ae066a1](https://github.com/jgsteeler/divergent-flow-core/commit/ae066a10231c41f0bebfe0294b407e7fd4136a59))
+
+### [0.1.9](https://github.com/jgsteeler/divergent-flow-core/compare/v0.1.8...v0.1.9) (2025-10-12)
+
+
+### Features
+
+* add development API service and update Docker commands in package.json ([c31517c](https://github.com/jgsteeler/divergent-flow-core/commit/c31517c08197841e390dc33f05381107dde53dba))
+
+
+### Bug Fixes
+
+* resolve TypeScript ignoreDeprecations configuration for Docker build ([5558252](https://github.com/jgsteeler/divergent-flow-core/commit/555825234715566edfb60d77bade19841bf9cc83))
+* update build script to build packages in dependency order ([45c5964](https://github.com/jgsteeler/divergent-flow-core/commit/45c596495208bc32d7a13b13b902f37920812f9d))
+
 ### [0.1.8](https://github.com/jgsteeler/divergent-flow-core/compare/v0.1.7...v0.1.8) (2025-10-12)